+++ R/model_get_model_frame.R
@@ -58,6 +58,10 @@ model_get_model_frame.coxph <- function(model) {
 
 #' @export
 #' @rdname model_get_model_frame
+model_get_model_frame.svycoxph <- model_get_model_frame.default
+
+#' @export
+#' @rdname model_get_model_frame
 model_get_model_frame.survreg <- function(model) {
   tryCatch(
     stats::model.frame.default(model),