+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - variable labels are now returned by `model_list_variables()` for `svycoxph`
   models (#275)
+- compatibility with R version 4.1 minimum (#276)
 
 # broom.helpers 1.17.0